Moved location and permissions code to packages
