+++ app/src/main/java/com/example/campusguide/location/CenterLocationListener.kt
@@ -1,9 +1,13 @@
-package com.example.campusguide
+package com.example.campusguide.location
 
 import android.Manifest
 import android.app.Activity
 import android.location.Location
 import android.view.View
+import com.example.campusguide.Constants
+import com.example.campusguide.Map
+import com.example.campusguide.permissions.PermissionGrantedObserver
+import com.example.campusguide.permissions.Permissions
 import com.google.android.gms.location.FusedLocationProviderClient
 import com.google.android.gms.location.LocationServices
 import com.google.android.gms.maps.model.LatLng
@@ -12,7 +16,8 @@ class CenterLocationListener constructor(
     private val activity: Activity,
     private val map: Map,
     private val permissions: Permissions
-) : View.OnClickListener, PermissionGrantedObserver {
+) : View.OnClickListener,
+    PermissionGrantedObserver {
     init {
         permissions.addObserver(this)
     }
@@ -50,7 +55,9 @@ class CenterLocationListener constructor(
     private fun animateCurrentLocation(location: Location) {
         val currentLatLng = LatLng(location.latitude, location.longitude)
         map.addMarker(currentLatLng, "You are here.")
-        map.animateCamera(currentLatLng, Constants.ZOOM_STREET_LVL)
+        map.animateCamera(currentLatLng,
+            Constants.ZOOM_STREET_LVL
+        )
     }
 
 }
