+++ app/src/main/java/com/example/campusguide/permissions/PermissionGrantedObserver.kt
@@ -1,4 +1,4 @@
-package com.example.campusguide
+package com.example.campusguide.permissions
 
 interface PermissionGrantedObserver {
     fun onPermissionGranted(permission: String)
